% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druglikeness.R
\name{assess_drug_likeness}
\alias{assess_drug_likeness}
\title{Assess drug-likeness}
\usage{
assess_drug_likeness(properties_df)
}
\arguments{
\item{properties_df}{Data frame with molecular properties.}
}
\value{
Data frame with an additional column indicating whether each molecule is drug-like or not.
}
\description{
This function assesses the drug-likeness of molecules based on calculated molecular properties.
}
