% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{compute_properties}
\alias{compute_properties}
\title{Compute molecular properties}
\usage{
compute_properties(mols)
}
\arguments{
\item{mols}{Molecules data.}
}
\value{
Data frame with calculated properties including Molecular Weight (MW),
number of Hydrogen Bond Donors (nHBDon), number of Hydrogen Bond Acceptors (nHBAcc),
Topological Polar Surface Area (TPSA), and ALogP.
}
\description{
This function computes molecular properties based on input molecules.
}
