% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_lipinski_plots}
\alias{create_lipinski_plots}
\title{Create Lipinski plots}
\usage{
create_lipinski_plots(properties_df)
}
\arguments{
\item{properties_df}{Data frame with molecular properties and pass/fail status.}
}
\value{
A grid of plots showing distributions and pass/fail status.
}
\description{
This function creates plots illustrating the distribution of molecular properties
and the pass/fail status of Lipinski's Rule of Five.
}
