% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability.R
\name{predict_oral_bioavailability}
\alias{predict_oral_bioavailability}
\title{Predict oral bioavailability}
\usage{
predict_oral_bioavailability(properties_df)
}
\arguments{
\item{properties_df}{Data frame with molecular properties.}
}
\value{
Data frame with an additional column indicating whether each molecule has high or low oral bioavailability.
}
\description{
This function predicts the oral bioavailability of molecules based on calculated molecular properties.
}
