% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_lipinski_rule}
\alias{test_lipinski_rule}
\title{Test Lipinski's Rule of Five}
\usage{
test_lipinski_rule(properties_df)
}
\arguments{
\item{properties_df}{Data frame with molecular properties.}
}
\value{
A vector indicating whether each molecule passes or fails Lipinski's Rule of Five.
}
\description{
This function tests Lipinski's Rule of Five based on calculated molecular properties.
}
